refactor(cardbuilder/library view): apply suggestions from review
